refactor: also extract common logic for SVMs
