+++ src/safeds/ml/classical/regression/_support_vector_regressor.py
@@ -1,22 +1,19 @@
 from __future__ import annotations
 
-import sys
-from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 from safeds._utils import _structural_hash
-from safeds.exceptions import ClosedBound, OpenBound, OutOfBoundsError
+from safeds.ml.classical._bases import _SupportVectorMachineBase
 from safeds.ml.classical.regression import Regressor
 
 if TYPE_CHECKING:
     from sklearn.base import RegressorMixin
-    from sklearn.svm import SVR as SklearnSVR  # noqa: N811
 
     from safeds.data.labeled.containers import TabularDataset
     from safeds.data.tabular.containers import Table
 
 
-class SupportVectorRegressor(Regressor):
+class SupportVectorRegressor(Regressor, _SupportVectorMachineBase):
     """
     Support vector machine for classification.
 
@@ -34,58 +31,6 @@ class SupportVectorRegressor(Regressor):
     """
 
     # ------------------------------------------------------------------------------------------------------------------
-    # Inner classes
-    # ------------------------------------------------------------------------------------------------------------------
-
-    class Kernel(ABC):
-        """Possible kernels for the support vector machine. Use the inner classes to create instances of this class."""
-
-        @abstractmethod
-        def __eq__(self, other: object) -> bool: ...
-
-        @abstractmethod
-        def __hash__(self) -> int: ...
-
-        @abstractmethod
-        def __str__(self) -> str: ...
-
-        @abstractmethod
-        def _apply(self, model: SklearnSVR) -> None:
-            """Set the kernel of the given model."""
-
-        @staticmethod
-        def Linear() -> SupportVectorRegressor.Kernel:  # noqa: N802
-            """A linear kernel."""  # noqa: D401
-            raise NotImplementedError  # pragma: no cover
-
-        @staticmethod
-        def Polynomial(degree: int) -> SupportVectorRegressor.Kernel:  # noqa: N802
-            """
-            A polynomial kernel.
-
-            Parameters
-            ----------
-            degree:
-                The degree of the polynomial kernel. Must be greater than 0.
-
-            Raises
-            ------
-            ValueError
-                If `degree` is not greater than 0.
-            """  # noqa: D401
-            raise NotImplementedError  # pragma: no cover
-
-        @staticmethod
-        def RadialBasisFunction() -> SupportVectorRegressor.Kernel:  # noqa: N802
-            """A radial basis function kernel."""  # noqa: D401
-            raise NotImplementedError  # pragma: no cover
-
-        @staticmethod
-        def Sigmoid() -> SupportVectorRegressor.Kernel:  # noqa: N802
-            """A sigmoid kernel."""  # noqa: D401
-            raise NotImplementedError  # pragma: no cover
-
-    # ------------------------------------------------------------------------------------------------------------------
     # Dunder methods
     # ------------------------------------------------------------------------------------------------------------------
 
@@ -95,52 +40,19 @@ class SupportVectorRegressor(Regressor):
         c: float = 1.0,
         kernel: SupportVectorRegressor.Kernel | None = None,
     ) -> None:
-        super().__init__()
-
-        # Inputs
-        if c <= 0:
-            raise OutOfBoundsError(c, name="c", lower_bound=OpenBound(0))
-        if kernel is None:
-            kernel = SupportVectorRegressor.Kernel.RadialBasisFunction()
-
-        # Hyperparameters
-        self._c: float = c
-        self._kernel: SupportVectorRegressor.Kernel = kernel
+        # Initialize superclasses
+        Regressor.__init__(self)
+        _SupportVectorMachineBase.__init__(
+            self,
+            c=c,
+            kernel=kernel,
+        )
 
     def __hash__(self) -> int:
         return _structural_hash(
             Regressor.__hash__(self),
-            self._c,
-            self.kernel,
+            _SupportVectorMachineBase.__hash__(self),
         )
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Properties
-    # ------------------------------------------------------------------------------------------------------------------
-
-    @property
-    def c(self) -> float:
-        """
-        Get the regularization strength.
-
-        Returns
-        -------
-        result:
-            The regularization strength.
-        """
-        return self._c
-
-    @property
-    def kernel(self) -> SupportVectorRegressor.Kernel:
-        """
-        Get the type of kernel used.
-
-        Returns
-        -------
-        result:
-            The type of kernel used.
-        """
-        return self._kernel
 
     # ------------------------------------------------------------------------------------------------------------------
     # Template methods
@@ -174,137 +86,3 @@ class SupportVectorRegressor(Regressor):
         )
         self._kernel._apply(result)
         return result
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Kernels
-# ----------------------------------------------------------------------------------------------------------------------
-
-class _Linear(SupportVectorRegressor.Kernel):
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Dunder methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, _Linear):
-            return NotImplemented
-        return True
-
-    def __hash__(self) -> int:
-        return _structural_hash(self.__class__.__qualname__)
-
-    def __str__(self) -> str:
-        return "Linear"
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Template methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def _apply(self, model: SklearnSVR) -> None:
-        model.kernel = "linear"
-
-
-class _Polynomial(SupportVectorRegressor.Kernel):
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Dunder methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def __init__(self, degree: int):
-        if degree < 1:
-            raise OutOfBoundsError(degree, name="degree", lower_bound=ClosedBound(1))
-
-        self._degree = degree
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, _Polynomial):
-            return NotImplemented
-        return self._degree == other._degree
-
-    def __hash__(self) -> int:
-        return _structural_hash(
-            self.__class__.__qualname__,
-            self._degree,
-        )
-
-    def __sizeof__(self) -> int:
-        return sys.getsizeof(self._degree)
-
-    def __str__(self) -> str:
-        return f"Polynomial(degree={self._degree})"
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Properties
-    # ------------------------------------------------------------------------------------------------------------------
-
-    @property
-    def degree(self) -> int:
-        """The degree of the polynomial kernel."""
-        return self._degree
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Template methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def _apply(self, model: SklearnSVR) -> None:
-        model.kernel = "poly"
-        model.degree = self._degree
-
-
-class _RadialBasisFunction(SupportVectorRegressor.Kernel):
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Dunder methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, _RadialBasisFunction):
-            return NotImplemented
-        return True
-
-    def __hash__(self) -> int:
-        return _structural_hash(self.__class__.__qualname__)
-
-    def __str__(self) -> str:
-        return "RadialBasisFunction"
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Template methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def _apply(self, model: SklearnSVR) -> None:
-        model.kernel = "rbf"
-
-
-class _Sigmoid(SupportVectorRegressor.Kernel):
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Dunder methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, _Sigmoid):
-            return NotImplemented
-        return True
-
-    def __hash__(self) -> int:
-        return _structural_hash(self.__class__.__qualname__)
-
-    def __str__(self) -> str:
-        return "Sigmoid"
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Template methods
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def _apply(self, model: SklearnSVR) -> None:
-        model.kernel = "sigmoid"
-
-
-# Override the methods with classes, so they can be used in `isinstance` calls. Unlike methods, classes define a type.
-# This is needed for the DSL, where SVM kernels are variants of an enum.
-SupportVectorRegressor.Kernel.Linear = _Linear  # type: ignore[method-assign]
-SupportVectorRegressor.Kernel.Polynomial = _Polynomial  # type: ignore[method-assign]
-SupportVectorRegressor.Kernel.RadialBasisFunction = _RadialBasisFunction  # type: ignore[method-assign]
-SupportVectorRegressor.Kernel.Sigmoid = _Sigmoid  # type: ignore[method-assign]
